EPMRPP-38754. Replace unassign dialog by common confirmation dialog.
