+++ app/src/pages/common/membersPage/index.js
@@ -2,4 +2,3 @@ export { MembersPage } from './membersPage';
 export { PermissionMapModal } from './modals/permissionMapModal';
 export { InviteUserModal } from './modals/inviteUserModal';
 export { ExternalUserInvitationModal } from './modals/externalUserInvitationModal';
-export { UnassignModal } from './modals/unassignModal';
