--- app/src/pages/common/membersPage/modals/unassignModal/index.js
@@ -1,1 +1,0 @@
-export { UnassignModal } from './unassignModal';
