--- app/src/pages/common/membersPage/modals/unassignModal/unassignModal.scss
@@ -1,6 +1,0 @@
-.message {
-  line-height: 22px;
-  b {
-    font-family: $FONT-SEMIBOLD;
-  }
-}
