+++ app/src/pages/common/modals/confirmationModal.scss
@@ -1,0 +1,10 @@
+.message {
+  line-height: 22px;
+  word-break: break-all;
+  @supports(word-break: break-word) {
+    word-break: break-word;
+  }
+  b {
+    font-family: $FONT-SEMIBOLD;
+  }
+}
